add updateFunction in image Reop and then change updateFunction in user and product Controller, run ok 3 updates

+++ Controllers/ProductController.cs
@@ -26,10 +26,10 @@ namespace shoppingapi2.Controllers
         {
             var product = await _productRepository.AddAsync(addProductDto);
             var response = _mapper.Map<AdminProductResponseDto>(product);
-            List<Image> imgLst=new();
-            foreach(var item in addProductDto.Files )
+            List<Image> imgLst = new();
+            foreach (var item in addProductDto.Files)
             {
-               imgLst.Add(await _imageRepository.SaveAsync(item, "Product", product.Id));
+                imgLst.Add(await _imageRepository.SaveAsync(item, "Product", product.Id));
             }
             response.Images = imgLst;
             return Ok(response);
@@ -44,7 +44,7 @@ namespace shoppingapi2.Controllers
                 var responses = products?.Select(x => _mapper.Map<AdminProductResponseDto>(x)).ToList();
                 foreach (var r in responses!)
                 {
-                    r.Images=await _imageRepository.GetImagesAsync("Product",r.Id);
+                    r.Images = await _imageRepository.GetImagesAsync("Product", r.Id);
                 }
                 return Ok(responses);
             }
@@ -56,10 +56,10 @@ namespace shoppingapi2.Controllers
         public async Task<IActionResult> GetByIdAsync(int id)
         {
             var product = await _productRepository.GetByIdAsync(id);
-            if(product != null)
+            if (product != null)
             {
-                var response=_mapper.Map<AdminProductResponseDto>(product);
-                response.Images=await _imageRepository.GetImagesAsync("Product",product.Id);
+                var response = _mapper.Map<AdminProductResponseDto>(product);
+                response.Images = await _imageRepository.GetImagesAsync("Product", product.Id);
                 return Ok(response);
             }
             return Ok($"product by {id} is not found!!");
@@ -67,32 +67,33 @@ namespace shoppingapi2.Controllers
         }
         [HttpPut]
         [Route("Update/{id}")]
-        public async Task<IActionResult> UpdateAsync(int id, UpdateProductDto updateProductDto)
+        public async Task<IActionResult> UpdateAsync(int id, [FromForm] UpdateProductDto updateProductDto)
         {
-            var result = await _productRepository.UpdateAsync(id, updateProductDto);
-            return Ok(result == true ? $"{id} number is updated" : $"{id}number is not found!!");
+            var productResult = await _productRepository.UpdateAsync(id, updateProductDto);
+            if (!productResult)
+                return Ok($"{id} number is not found!!");
+            //update images for this product
+            List<bool> imageResult = new();
+            foreach (var file in updateProductDto.Files)
+                imageResult.Add(await _imageRepository.UpdateAsync(file, "Product", id));
+            if (imageResult.All(x => x == true))
+                return Ok($"successfully updated the product and it's image");
+            else
+                return Ok($"Failed to update Images!");
         }
         [HttpDelete]
         [Route("Delete/{id}")]
         public async Task<IActionResult> DeleteAsync(int id)
         {
-            var resultproduct = await _productRepository.DeleteAsync(id);
-            var rsltImg=await _imageRepository.DeleteAsync("Product",id);
-            if (rsltImg==true) {
-                if(resultproduct==true)
-                    return Ok("Successfully both of deleted");
-                else 
-                return Ok("Falied product, succssfully images");
-
-            } 
+            var productResult = await _productRepository.DeleteAsync(id);
+            if (!productResult)
+                return Ok($"{id}number is not found!!");
+            var imageResult = await _imageRepository.DeleteAsync("Product", id);
+            if (imageResult)
+                return Ok("Successfuly deleted the product and it's image");
             else
-            {
-                if(resultproduct==true)
-                    return Ok("Falied images , succssfully product");
-                else 
-                return Ok("Falied both of them");
+                return Ok("Failed to delete Image");
 
-            }
         }
     }
 }
